Fixed Firebase Cleanup Error on Add Admin Screen
