+++ src/screens/Forum/ForumManagement/AddAdminScreen.js
@@ -2,24 +2,12 @@ import { useNavigation } from "@react-navigation/native";
 import { useState } from "react";
 import styled from "styled-components/native";
 import GetUserWithUsername from "../../../services/Friend/GetUserWithUsername";
-import { isUserBanned } from "../../../services/Forum/HandleBannedUsers";
-import { isUserAdmin } from "../../../services/Forum/HandleForumAdmin";
 import RenderUserToManage from "../../../components/Forum/ForumManagement/RenderUserToManage";
-
-const initialData = {
-  isBanned: false,
-  reason: "",
-  isAdmin: false,
-  authorities: []
-};
 
 const AddAdminScreen = () => {
   const [username, setUsername] = useState("");
   const [showUserDetails, setShowUserDetails] = useState(false);
-  const [data, setData] = useState(initialData); // For the user forum info
   const [userData, setUserData] = useState(null);
-  const navigation = useNavigation();
-  const forumId = navigation.getState().routes[1].params.data.id;
 
   function handleChangeText(text) {
     setUsername(text);
@@ -33,24 +21,6 @@ const AddAdminScreen = () => {
       specifiedUsername: username,
       onFound: (data) => {
         setUserData(data);
-        isUserBanned(forumId, data.id, (result) => {
-          setData((data) => {
-            return {
-              ...data,
-              isBanned: result.isFound,
-              reason: result.isFound ? result.reason : ""
-            };
-          });
-        });
-        isUserAdmin(forumId, data.id, (result) => {
-          setData((data) => {
-            return {
-              ...data,
-              isAdmin: result.isFound,
-              authorities: result.isFound ? result.authorities : []
-            };
-          });
-        });
       },
       onNotFound: () => {
         setUserData(null);
@@ -71,9 +41,7 @@ const AddAdminScreen = () => {
           <ButtonText> Search </ButtonText>
         </CustomButton>
       </SearchContainer>
-      {showUserDetails && (
-        <RenderUserToManage userData={userData} data={data} />
-      )}
+      {showUserDetails && <RenderUserToManage userData={userData} />}
     </Container>
   );
 };
